purego: add //go:uintptrescapes to SyscallN
this forces the uintptr arguments to escape in SyscallN. When optimizations were enabled Go assumed that the pointer arguments in SyscallN were no longer needed. This is why functions like gl.GenFramebuffersEXT would fail. Although this change forces all of the arguments to be placed on the heap it ensures that no optimizations will remove our precious arguments. This also allows us to only have one implementation inside SyscallN (i.e. syscall_syscall9X). A potential replacement would be //go:cgo_unsafe_args but that can only be used in CGO generated files. There is also //go:uintptrkeepalive although that was not enough to stop the function from failing. That is why I went with //go:uintptrescapes

I tested this change on darwin (arm64 & amd64) as well as with -gcflags=all=-N and without. All tests pass on these platforms.

I have not tested this on windows. If you could that would be much appreciated.