internal/fakecgo: support on macOS arm64 and amd64 (#20)
This commit reimplements the Cgo runtime found at runtime/cgo entirely in Go. It does this by using //go:linkname to replace the symbols in the Go runtime with assembly trampolines into Go functions. The symbols that are replaced are:

 iscgo             bool                                                  // runtime/cgo.go
_cgo_init         unsafe.Pointer
_cgo_thread_start unsafe.Pointer
_cgo_notify_runtime_init_done unsafe.Pointer
_cgo_setenv unsafe.Pointer                                 // these two are in runtime/env_posix.go
_cgo_unsetenv unsafe.Pointer

This commit also changes which trampoline is used to move from Go to C. It changes from runtime.libcCall to runtime.cgocall which is the function used when calling C functions with import "C". Doing this also brings purego more in line with Cgo. This will allow us to more easily add platforms and remove C from Go programs.

This commit adds support for fakecgo on macOS (arm64 and amd64). In the future, support for linux and iOS can be added with this commit as the basis for adding platforms. The platform specific parts include symbols_GOOS.go and go_GOOS_GOARCH.go. Implementing those for other GOOSs and GOARCHs should be trivial.

The current limitations is that fakecgo does no work on iOS nor any other GOOS but macOS.
It can also not be used with -buildmode=c-archive as that requires extra initialization that isn't done in this commit.

Updates #12

Co-authored-by: Hajime Hoshi <hajimehoshi@gmail.com>