purego: remove go:nosplit at SyscallN (#24)
this is allowed because SyscallN also uses go:uintptrescapes which forces any objects that get passed as uintptr's to be placed on the heap where a stack split won't affect them. See the compile docs for an explanation.

I've tested this change on Oto (amd64 & arm64) without optimizations. I also ran all the test for ebitengine (amd64 & arm64) with OpenGL and Metal.

All this to say, this won't break anything.

This change fixes Oto#182