fix callback calling convention (#26)
runtime.cgocallback takes a function pointer to the ABIInternal entry. The only way to get the ABIInternal entry is using the ABI selector <ABIInternal>. However, this is only available when compiling the runtime.  This commit circumvents this limitation by using build constraints to determine if the Go version uses stack or register calling convention. Then in callbackWrapPicker the proper argument is sent to the real callbackWrap function.

Closes #25