rename libc_* to purego_*
While porting purego to Linux I discovered that I had missed the cgo_import_dynamic comment for pthread_sigmask. The reason purego was working was because the symbol is normally linked to in the runtime. I decided to make a unique name for the imported symbol so that we aren't relying on runtime behavior that can easily change.