objc: add RegisterClass API (#59)
adds RegisterClass function which takes a Selector interface that it uses to add methods based on the argument sent to RegisterClass. It returns an Objective-C class that matches the pointer to struct or an error if something failed. I also included an example under example/objc/main_darwin.go.

* AddMethod & objc tag support Protocol

This adds Selector interface which allows to convert Go method names to Objc methods. There is a handwritten parser for the objc tag which matches that of an Objc Class definition. This adds support for adding Protocols from the tag.