use R10 instead of R12
on x86-64 R12 is callee-saved which means if we use it in callbackasm1 we MUST preserve its value. We were not doing that so it caused the SIGSEGV to happen in Oto. I replaced R12 with R10 because R10 is caller-saved so we don't need to worry about throwing away the value inside of it.