purego: SyscallN r2 return float
We originally discussed this change and decided against doing it because r2 usually returns the upper 64bits on 32bit platforms. However, on Windows amd64 r2 is actually the float return value.

https://github.com/golang/go/blob/7717ac151ae1556541dddc6a817ac04733f1af44/src/runtime/sys_windows_amd64.s#L66

And there is a TODO for arm64 as well.

https://github.com/golang/go/blob/2cea6cdb6016708e89c9472cdb0504731699681d/src/runtime/sys_windows_arm64.s#L91

Doing this makes supporting float returns easier without having to implement a full C calling convention for RegisterFunc