internal/fakecgo: circumvent Go 1.20's race code (#73)
Go 1.20 rewrote the race detector on darwin to be entirely Go. This means
that fakecgo had race functions placed inside of it which made it break.
This PR circumvents the race functions by adding //go:norace and using
closures to call the ABIInternal ABIWrapper which does not have the race
instrumentation.

Closes #71