use crosscall2 from runtime/cgo
The code originally inlined crosscall2 because at the time there was no internal/fakecgo package to place the implementation. Now that this package exists crosscall2 can be put into it. The implementation is copied from runtime/cgo. When CGO_ENABLED=1 it will use the runtime/cgo implementation of crosscall2 which should always be identical.

Less assembly to maintain is good! :D