better error message and padding support (#83)
This commit improves the error message when the Ivar and struct field are not at the same offset. 

If the parent class has fields its fields are located before the child's in the struct. This causes issues since there is no way to know the internal structure of the class so padding is required. This commit enables the use of padding by using the underscore (_) identifier with an array of bytes that matches the parent's field size. Still, if any fields are misaligned an error is produced.