internal/fakecgo: align stack on amd64 for SSE (#111)
Some SSE instructions on amd64 require that the stack be aligned. The stack was not properly aligned and would SIGSEGV inside fakecgo. This patch aligns the stack properly. It also merges the darwin and linux trampolines. This is a reverse of #104 which I believe was actually caused by this issue. It also explicitly tests the behavior with and without Cgo (CGO_ENABLED=0/1).

Closes #106