purego: add float32 callback test
This fixes a bug in RegisterFunc which was converting float32 into float64 when the C function expected a float32 argument. This gave unpredictable results. It also makes the behavior of SyscallN more inline with what is expected when calling it with float arguments. It no longer expects them to be float64 and it is up to the caller to convert either float32 or float64 into uintptr. SyscallN will then properly move it into the correct register