purego: add float32 and float64 support to callbacks (#120)
This commit adds support for float32 and float64 arguments in C callbacks. It also adds a test to check the maximum number of float32, float64 and mixed arguments a callback can have and that it is callback properly by purego. There is also some formatting done to the assembly.

It also fixes a bug in RegisterFunc which was converting float32 into float64 when the C function expected a float32 argument. This gave unpredictable results. It also makes the behavior of SyscallN more inline with what is expected when calling it with float arguments. It no longer expects them to be float64 and it is up to the caller to convert either float32 or float64 into uintptr using the math package. SyscallN will then properly move it into the correct register