purego: enable testing of amd64 with Rosseta on arm64 macOS (#122)
With Rosetta on arm64 it was possible to test amd64 just by setting GOARCH=amd64. However, since the shared library tests were added it is no longer possible because the default compiler will build for the host architecture and not the cross-compiled (amd64) architecture. This commit tells the C/C++ compiler to build for amd64 when running an amd64 binary. This only effects macOS.