Add FreeBSD text matrix with various Go binaries
The FreeBSD text needs to run in its own space because it's executed as
a virtualbox VM in macOS. vmactions/freebsd-vm only supports a single
step block where it exits the VM process once the unit complets. All the
run commands therefore need to be clumped together.

Different versions of Go are downloaded directly from the Go repo
instead of using FreeBSD's pkg, which is limited to only a single
version of Go.