all: add FreeBSD support (#155)
This change adds FreeBSD AMD64 and ARM64 targets.

Includes:

* github actions workflow with FreeBSD test unit:
The FreeBSD test needs to run in its own space because it's
executed as a virtualbox VM in macOS. vmactions/freebsd-vm only
supports a single step block where it exits the VM process once
the unit completse. All the run commands therefore need to be
clumped together.

The setup-go library doesn't support FreeBSD, so different
versions of Go are manually fetched directly from the Go repo
instead of using FreeBSD's pkg, which is limited to only a single
version of Go.

* environ and __progname variables to enable cross-compilation
which libc.so relies on. Cross-compile requires
`-gcflags="github.com/ebitengine/purego/internal/fakecgo=-std"`
to make the the symbols appear in the binary.

Sponsored by: Rubicon Communications LLC (Netgate)