internal/fakecgo: move go generate pragma into gen.go
The `go generate` command adds the `generate` build tag when searching for pragmas to run. This means the pragma doesn't need to be in a separate file.

This also means the documentation can be seen even on unsupported platforms as well as generating the files.

I also moved formatting into the code itself using the `go/format` package.