internal/fakecgo: preserve callee-saved registers in threadentry_trampoline (#347)
`threadentry_trampoline` is called from glibc (that is, C code) when
creating a new thread. When the thread exists, `threadentry_trampoline`
return back to glibc, which can perform some additional cleanups.

We currently aren't honoring the Linux C ABI. It specify some registers
as calle-saved which are clobbered by the Go code. This can cause
unexpected behaviors when returning to glibc when the thread exists,
as the C compiler assumes some registers will have the same value as
before the `threadentry_trampoline` function was called, but they don't.

This is a precondition to merge #344.
Also, might fix #304.