internal/fakecgo: don't overwrite the caller stack on arm64 trampolines (#379)
amd64 and arm64 assembly trampolines moved from stack arguments to register arguments in #73.
This means that we don't need to store the registers on the stack anymore. Note that doing so might
even lead to some nasty bugs, as we are writing on stack slots that haven't been allocated in the function
frame, potentially corrupting the caller stack.

Updates #343